test: migrate GameServiceTest, WebSocketBrokerControllerTest, DisconnectCleanupServiceTest to domain services
GameService wird in diesen Testdateien nicht mehr direkt verwendet.
